+++ src/test/java/dinkplugin/notifiers/ChatNotifierTest.java
@@ -7,10 +7,14 @@ import dinkplugin.message.NotificationType;
 import dinkplugin.message.templating.Template;
 import dinkplugin.notifiers.data.ChatNotificationData;
 import net.runelite.api.ChatMessageType;
+import net.runelite.api.events.CommandExecuted;
+import net.runelite.client.config.Notification;
+import net.runelite.client.events.NotificationFired;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 
+import java.awt.TrayIcon;
 import java.util.EnumSet;
 
 import static org.mockito.ArgumentMatchers.any;
@@ -32,10 +36,12 @@ public class ChatNotifierTest extends MockedNotifierTest {
 
         // config mocks
         when(config.notifyChat()).thenReturn(true);
-        when(config.chatMessageTypes()).thenReturn(EnumSet.of(ChatNotificationType.GAME));
+        when(config.chatMessageTypes()).thenReturn(EnumSet.of(ChatNotificationType.GAME, ChatNotificationType.COMMAND, ChatNotificationType.RUNELITE));
         when(config.chatNotifyMessage()).thenReturn("%USERNAME% received a chat message:\n\n```\n%MESSAGE%\n```");
         setPatterns("You will be logged out in approximately 10 minutes.*\n" +
-            "You will be logged out in approximately 5 minutes.*\n");
+            "You will be logged out in approximately 5 minutes.*\n" +
+            "Dragon impling is in the area\n" +
+            "::TriggerDink\n");
     }
 
     @Test
@@ -62,12 +68,61 @@ public class ChatNotifierTest extends MockedNotifierTest {
     }
 
     @Test
+    void testNotifyCommand() {
+        // fire event
+        String message = "::TriggerDink";
+        notifier.onCommand(new CommandExecuted(message.substring(2), new String[0]));
+
+        // verify notification message
+        verify(messageHandler).createMessage(
+            PRIMARY_WEBHOOK_URL,
+            false,
+            NotificationBody.builder()
+                .text(
+                    Template.builder()
+                        .template(PLAYER_NAME + " received a chat message:\n\n```\n" + message + "\n```")
+                        .build()
+                )
+                .extra(new ChatNotificationData(ChatMessageType.UNKNOWN, "CommandExecuted", message))
+                .type(NotificationType.CHAT)
+                .playerName(PLAYER_NAME)
+                .build()
+        );
+    }
+
+    @Test
+    void testNotifyTray() {
+        // fire event
+        String message = "Dragon impling is in the area";
+        notifier.onNotification(new NotificationFired(Notification.ON, message, TrayIcon.MessageType.INFO));
+
+        // verify notification message
+        verify(messageHandler).createMessage(
+            PRIMARY_WEBHOOK_URL,
+            false,
+            NotificationBody.builder()
+                .text(
+                    Template.builder()
+                        .template(PLAYER_NAME + " received a chat message:\n\n```\n" + message + "\n```")
+                        .build()
+                )
+                .extra(new ChatNotificationData(ChatMessageType.UNKNOWN, "NotificationFired", message))
+                .type(NotificationType.CHAT)
+                .playerName(PLAYER_NAME)
+                .build()
+        );
+    }
+
+    @Test
     void testIgnore() {
         // fire event
         notifier.onMessage(ChatMessageType.GAMEMESSAGE, null, "You will be logged out in approximately 30 minutes.");
         notifier.onMessage(ChatMessageType.PUBLICCHAT, null, "You will be logged out in approximately 10 minutes.");
         notifier.onMessage(ChatMessageType.TRADE, null, "You will be logged out in approximately 10 minutes.");
         notifier.onMessage(ChatMessageType.CLAN_MESSAGE, null, "You will be logged out in approximately 10 minutes.");
+        notifier.onCommand(new CommandExecuted("You", "will be logged out in approximately 10 minutes.".split(" ")));
+        notifier.onCommand(new CommandExecuted("DontTriggerDink", new String[0]));
+        notifier.onNotification(new NotificationFired(Notification.ON, "TriggerDink", TrayIcon.MessageType.INFO));
 
         // ensure no notification occurred
         verify(messageHandler, never()).createMessage(any(), anyBoolean(), any());
@@ -80,6 +135,7 @@ public class ChatNotifierTest extends MockedNotifierTest {
 
         // fire event
         notifier.onMessage(ChatMessageType.GAMEMESSAGE, null, "You will be logged out in approximately 10 minutes.");
+        notifier.onCommand(new CommandExecuted("TriggerDink", new String[0]));
 
         // ensure no notification occurred
         verify(messageHandler, never()).createMessage(any(), anyBoolean(), any());